fix LISPSM warping frustum computation
We were using the view frustum as a basis instead of the shadow 
receivers volume which didn't give a good warping for the scene, we
were working around it by setting a virtual near plane dynamically.

Now that the warping frustum is tight, the virtual near plane is 
no longer needed -- it's still needed for the user to adjust the
precision and the default is still 1m.